# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Kenya Location Intelligence API",
    "description": "Geospatial API for Kenya administrative boundaries, places, nearby searches, and reverse geolocation.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/places/search": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Search places",
        "description": "Searches places by name or place type.",
        "operationId": "SearchPlaces",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Optional text used to search by place name.",
            "schema": {
              "type": "string",
              "description": "Optional text used to search by place name."
            }
          },
          {
            "name": "placeType",
            "in": "query",
            "description": "Optional place type. Supported values: urban_area, settlement.",
            "schema": {
              "type": "string",
              "description": "Optional place type. Supported values: urban_area, settlement."
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
            "schema": {
              "type": "integer",
              "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceSearchResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/{id}": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Get a place",
        "description": "Returns detailed information and GeoJSON geometry for a place.",
        "operationId": "GetPlace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Place identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Place identifier."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiNotFoundError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/nearby": {
      "get": {
        "tags": [
          "Places"
        ],
        "summary": "Find nearby places",
        "description": "Finds places near a latitude and longitude using a radius in kilometres.",
        "operationId": "GetNearbyPlaces",
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "description": "Latitude in decimal degrees. Valid range: -90 to 90.",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Latitude in decimal degrees. Valid range: -90 to 90.",
              "format": "double"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "description": "Longitude in decimal degrees. Valid range: -180 to 180.",
            "required": true,
            "schema": {
              "type": "number",
              "description": "Longitude in decimal degrees. Valid range: -180 to 180.",
              "format": "double"
            }
          },
          {
            "name": "radiusKm",
            "in": "query",
            "description": "Search radius in kilometres. Default: 25. Valid range: 0.1 to 500.",
            "schema": {
              "type": "number",
              "description": "Search radius in kilometres. Default: 25. Valid range: 0.1 to 500.",
              "format": "double"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
            "schema": {
              "type": "integer",
              "description": "Maximum number of results to return. Default: 20. Valid range: 1 to 100.",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NearbyPlaceResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApiNotFoundError": {
        "required": [
          "error",
          "id"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "NearbyPlaceResponse": {
        "required": [
          "distanceMeters",
          "id",
          "latitude",
          "longitude",
          "name",
          "placeType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "placeType": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "distanceMeters": {
            "type": "number",
            "format": "double"
          },
          "ward": {
            "type": "string",
            "nullable": true
          },
          "subCounty": {
            "type": "string",
            "nullable": true
          },
          "county": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlaceDetailResponse": {
        "required": [
          "geometry",
          "id",
          "latitude",
          "longitude",
          "name",
          "placeType",
          "source",
          "sourceVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "placeType": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "geometry": { },
          "wardId": {
            "type": "string",
            "nullable": true
          },
          "ward": {
            "type": "string",
            "nullable": true
          },
          "subCountyId": {
            "type": "string",
            "nullable": true
          },
          "subCounty": {
            "type": "string",
            "nullable": true
          },
          "countyId": {
            "type": "string",
            "nullable": true
          },
          "county": {
            "type": "string",
            "nullable": true
          },
          "adminUnitId": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string"
          },
          "sourceVersion": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PlaceSearchResponse": {
        "required": [
          "id",
          "latitude",
          "longitude",
          "name",
          "placeType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "placeType": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "ward": {
            "type": "string",
            "nullable": true
          },
          "subCounty": {
            "type": "string",
            "nullable": true
          },
          "county": {
            "type": "string",
            "nullable": true
          },
          "adminUnitId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}